+++ src/parakeet/Pro/Driver.cpp
@@ -4,7 +4,6 @@
 
 #include <parakeet/Pro/Driver.h>
 
-#include <parakeet/exceptions/NotConnectedToSensorException.h>
 #include <parakeet/exceptions/UnableToDetermineBaudRateException.h>
 #include <parakeet/exceptions/UnableToOpenPortException.h>
 
@@ -14,23 +13,12 @@ namespace parakeet
 {
 namespace Pro
 {
-    const int MAX_NUMBER_OF_POINTS_FROM_SENSOR = 1000;  // Arbitrary size
-    const int MESSAGE_DATA_BUFFER_SIZE = 8192;          // Arbitrary size
-
-    struct Driver::ScanData
-    {
-        unsigned short from;
-        unsigned short span;
-        unsigned short count;
-        unsigned short reserved;
-        unsigned short dist[MAX_NUMBER_OF_POINTS_FROM_SENSOR];
-        unsigned char intensity[MAX_NUMBER_OF_POINTS_FROM_SENSOR];
-    };
+    const int SERIAL_MESSAGE_DATA_BUFFER_SIZE = 8192;          // Arbitrary size
 
     struct Driver::MessageData
     {
         int len;
-        unsigned char body[MESSAGE_DATA_BUFFER_SIZE];
+        unsigned char body[SERIAL_MESSAGE_DATA_BUFFER_SIZE];
     };
 
     const std::string CW_STOP_ROTATING = "LSTOPH";
@@ -70,12 +58,14 @@ namespace Pro
     {
         return SW_SET_BIAS_PREFIX + std::to_string(bias) + SW_SET_BIAS_POSTFIX;
     }
-
-    Driver::~Driver()
-    {
-        close();
-    }
     
+    Driver::Driver()
+    {
+        this->registerUpdateThreadCallback(std::bind(&Driver::serialUpdateThreadFunction, this));
+
+        serialPortDataBuffer = new unsigned char[SERIAL_MESSAGE_DATA_BUFFER_SIZE];
+    }
+
     void Driver::connect(const SensorConfiguration& sensorConfiguration)
     {
         this->sensorConfiguration = sensorConfiguration;
@@ -147,12 +137,9 @@ namespace Pro
 
     void Driver::start()
     {
-        throwExceptionIfNotConnected();
+        serialPortDataBufferLength = 0;
 
-        runSerialInterfaceThread = true;
-        interfaceThreadStartTime = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch());
-        interfaceThreadFrameCount = 0;
-        serialInterfaceThread = std::thread([&] { this->serialInterfaceThreadFunction(); });
+        mechaspin::parakeet::Driver::start();
 
         sendMessageWaitForResponseOrTimeout(internal::SensorResponse::START, CW_START_NORMALLY, std::chrono::milliseconds(1000));
 
@@ -166,27 +153,14 @@ namespace Pro
     {
         serialPort.write(CW_STOP_ROTATING);
 
-        runSerialInterfaceThread = false;
-
-        if(serialInterfaceThread.joinable())
-        {
-            serialInterfaceThread.join();
-        }
+        mechaspin::parakeet::Driver::stop();
     }
 
     void Driver::close()
     {
-        if (serialPort.isConnected() || runSerialInterfaceThread || serialInterfaceThread.joinable())
-        {
-           stop();
-        }
+        mechaspin::parakeet::Driver::close();
 
         serialPort.close();
-    }
-
-    void Driver::registerScanCallback(std::function<void(const ScanDataPolar&)> callback)
-    {
-        scanCallbackFunction = callback;
     }
 
     void Driver::enableDataSmoothing(bool enable)
@@ -242,7 +216,7 @@ namespace Pro
 
         sendMessageWaitForResponseOrTimeout(internal::SensorResponse::BAUDRATE, SW_SET_BAUD_RATE(baudRate.getValue()), std::chrono::milliseconds(0));
 
-        if(runSerialInterfaceThread)
+        if(isUpdateThreadRunning())
         {
             stop();
 
@@ -286,50 +260,30 @@ namespace Pro
         return sensorConfiguration.scanningFrequency_Hz;
     }
 
-    double Driver::getScanRate_Hz()
+    void Driver::serialUpdateThreadFunction()
     {
-        std::chrono::milliseconds currentTime = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch());
-
-        auto ms = (currentTime - interfaceThreadStartTime).count();
-
-        if(ms == 0)
+        // modifying the baud rate of serial port can cause the connected state to be off for a brief moment
+        if (!serialPort.isConnected())
         {
-            return 0;
+            return;
         }
 
-        return interfaceThreadFrameCount / (ms / 1000.0);
-    }
+        int charsRead = serialPort.read(serialPortDataBuffer, serialPortDataBufferLength, SERIAL_MESSAGE_DATA_BUFFER_SIZE);
 
-    void Driver::serialInterfaceThreadFunction()
-    {
-        unsigned char* line = new unsigned char[MESSAGE_DATA_BUFFER_SIZE];
-        int len = 0;
-
-        while (runSerialInterfaceThread)
+        if (charsRead == 0)
         {
-            // modifying the baud rate of serial port can cause the connected state to be off for a brief moment
-            if (!serialPort.isConnected())
-            {
-                continue;
-            }
-
-            int charsRead = serialPort.read(line, len, MESSAGE_DATA_BUFFER_SIZE);
-
-            if (charsRead == 0)
-            {
-                continue;
-            }
-
-            len += charsRead;
-
-            int nl = parseSensorDataFromBuffer(len, line);
-
-            for (int i = nl; i<len; i++)
-            {
-                line[i - nl] = line[i];
-            }
-            len -= nl;
+            return;
         }
+
+        serialPortDataBufferLength += charsRead;
+
+        unsigned int nl = parseSensorDataFromBuffer(serialPortDataBufferLength, serialPortDataBuffer);
+
+        for (unsigned int i = nl; i < serialPortDataBufferLength; i++)
+        {
+            serialPortDataBuffer[i - nl] = serialPortDataBuffer[i];
+        }
+        serialPortDataBufferLength -= nl;
     }
 
     void Driver::onMessageDataReceived(MessageData* message)
@@ -344,43 +298,6 @@ namespace Pro
         }
 
         delete message;
-    }
-
-    void Driver::onScanDataReceived(ScanData* scanData)
-    {
-        if(pointHoldingList.size() == 0)
-        {
-            timeOfFirstPoint = std::chrono::system_clock::now();
-        }
-
-        double startAngle_deg = scanData->from / 10.0;
-        double endAngle_deg = (static_cast<double>(scanData->from) + static_cast<double>(scanData->span)) / 10.0;
-        double anglePerPoint_deg = (endAngle_deg - startAngle_deg) / scanData->count;
-        double deviationFrom360_deg = 1;
-
-        //Create PointPolar for each data point
-        for(int i = 0; i < scanData->count; i++)
-        {
-            PointPolar pointPolar(scanData->dist[i], startAngle_deg + (anglePerPoint_deg * i), scanData->intensity[i]);
-
-            pointHoldingList.push_back(pointPolar);
-        }
-
-        if(endAngle_deg + deviationFrom360_deg >= 360)
-        {
-            interfaceThreadFrameCount++;
-
-            ScanDataPolar scanDataPolar(pointHoldingList, timeOfFirstPoint);
-
-            if (scanCallbackFunction != nullptr)
-            {
-                scanCallbackFunction(scanDataPolar);
-            }
-
-            pointHoldingList.clear();
-        }
-
-        delete scanData;
     }
 
     int Driver::parseSensorDataFromBuffer(int length, unsigned char* buf)
@@ -563,12 +480,9 @@ namespace Pro
         return sensorReturnMessageState[messageType];
     }
 
-    void Driver::throwExceptionIfNotConnected()
+    bool Driver::isConnected()
     {
-        if(!serialPort.isConnected())
-        {
-            throw exceptions::NotConnectedToSensorException();
-        }
+        return serialPort.isConnected();
     }
 }
 }