+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - In Development
+### Added
+- Added a base driver class mechaspin::parakeet::Driver which holds common functionality between sensor drivers
+
+### Modified
+- [BREAKING] Replaced mechaspin::parakeet::Driver with mechaspin::parakeet::Pro::Driver
+
 ## [2.1.0] - 2021-07-13
 ### Deprecated
 - Deprecated class mechaspin::parakeet::Driver, replaced by mechaspin::parakeet::Pro::Driver